Generic parsing restructuring
This is a zero side-effect restructuring of clinto to prepare for
other parsers. Parsers are defined off a base class which provides
a series of stub functions that must be overridden to provide
a new parser.  The initialisation of a parser accepts a script
path and source code which is then tested, first for file extension
then for the file containing a magic token (e.g. ArgumentParser) that
identifies that file as likely to be using a particular parser.

If these match the parsing is performed and the parse object parsed(!)

If this final step is sucessful parsing will halt, otherwise it will
continue on to try the next parser.

Individual parsers are isolated in separate files (e.g. argparse.py)
and are free to use their own logic and internals to generate the output.
The json output of sucessfully processed parsers is available via the
public parser.Parser class (and is the get_script_description for debug).

This incorporates the changes for exception monkey patching.
